+++ changelog.md
@@ -5,4 +5,5 @@
     - accepting edits  
     - edit/print in sync with current browsing node  
     - print in live sync with edit  
+    - rudimentay image support